Moved wrapper module into its own package and initialized API object in skyAPI module
